refactor(core): move normalizeModelOutput to loop/shared
The main loop's tool-call-step imported normalizeModelOutput from
agent/durable — a layering violation where the shared loop layer
reached into a specific agent implementation. The helper is used by
both loops and by prepare-stream, so it belongs with the other shared
loop helpers in loop/shared. Pure move, no behavior change.

Note: llm-mapping-step.ts keeps its divergent inline copy for now;
merging that duplicate is a behavior decision deferred to the
chunk-pipeline hoist.
