+++ packages/core/src/loop/shared/normalize-model-output.ts
